P0_GS_1st page
 website should have at least three pages: one for regular Google Search (which must be called
index.html ), one for Google Image Search, and one for Google Advanced Search.

On the Google Search page, there should be links in the upper-right of the page to go to Image
Search or Advanced Search. On each of the other two pages, there should be a link in the
upper-right to go back to Google Search.